+++ src/screens/home/components/online_voting_power/index.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import classnames from 'classnames';
-import Link from 'next/link';
 import numeral from 'numeral';
 import { Typography } from '@material-ui/core';
 import useTranslation from 'next-translate/useTranslation';
 import { Box } from '@components';
-import { BLOCK_DETAILS } from '@utils/go_to_page';
 import { useStyles } from './styles';
 import { useOnlineVotingPower } from './hooks';
 
@@ -44,13 +42,11 @@ const OnlineVotingPower: React.FC<{
       <div className={classes.itemsContainer}>
         <div className={classes.item}>
           <Typography variant="h4" className="label">
-            {t('block')}
+            {t('validators')}
           </Typography>
-          <Link href={BLOCK_DETAILS(state.height)} passHref>
-            <Typography variant="body1" className="value" component="a">
-              {numeral(state.height).format('0,0')}
-            </Typography>
-          </Link>
+          <Typography variant="body1" className="value" component="a">
+            {numeral(state.activeValidators).format('0,0')}
+          </Typography>
         </div>
         <div className={classes.item}>
           <Typography variant="h4" className="label">